Fixed warning for ambiguous call

+++ src/DS1624.cpp
@@ -109,7 +109,7 @@ float DS1624::GetTemperature()
       // when conversion done
       Wire.beginTransmission(_address);
       Wire.write(ACCESS_CONFIG);
-      Wire.requestFrom(_address, 1);
+      Wire.requestFrom(_address, (size_t)1);
       
       // Wait for data sent from sensor
       while(!Wire.available());
@@ -122,7 +122,7 @@ float DS1624::GetTemperature()
   // Request to read last converted temperature value
   Wire.beginTransmission(_address);
   Wire.write(READ_TEMPERATURE);
-  Wire.requestFrom(_address, 2);
+  Wire.requestFrom(_address, (size_t)2);
 
   // Wait for data sent from sensor
   while(!Wire.available());
